refactor(components/molecule/modal): api tab
