+++ components/molecule/modal/src/index.js
@@ -1,11 +1,49 @@
-import MoleculeModal from './MoleculeModal'
-import MoleculeModalWithAnimation from './MoleculeModalWithAnimation'
-import MoleculeModalWithURLState from './MoleculeModalWithURLState'
-import MoleculeModalWithoutAnimation from './MoleculeModalWithoutAnimation'
+import {forwardRef, useState} from 'react'
+import PropTypes from 'prop-types'
+
 import {MODAL_SIZES} from './config'
+import MoleculeModal from './MoleculeModal'
+import MoleculeModalWithoutAnimation from './MoleculeModalWithoutAnimation'
+import MoleculeModalWithURLState from './MoleculeModalWithURLState'
 
 import MoleculeModalContent from './Content'
 import MoleculeModalFooter from './Footer'
+
+const MoleculeModalWithAnimation = forwardRef(
+  ({onClose, onAnimationEnd, ...rest}, ref) => {
+    const [isClosing, setIsClosing] = useState(false)
+
+    const handleAnimationEnd = event => {
+      typeof onAnimationEnd === 'function' && onAnimationEnd()
+
+      if (!isClosing) return
+
+      setIsClosing(false)
+      typeof onClose === 'function' && onClose(event)
+    }
+
+    const handleClose = () => {
+      setIsClosing(true)
+    }
+
+    return (
+      <MoleculeModal
+        ref={ref}
+        isClosing={isClosing}
+        onAnimationEnd={handleAnimationEnd}
+        onClose={handleClose}
+        {...rest}
+      />
+    )
+  }
+)
+
+MoleculeModalWithAnimation.displayName = `(${MoleculeModal.displayName})WithAnimation`
+MoleculeModalWithAnimation.contextTypes = MoleculeModal.contextTypes
+MoleculeModalWithAnimation.propTypes = {
+  onClose: PropTypes.func,
+  onAnimationEnd: PropTypes.func
+}
 
 MoleculeModalWithAnimation.displayName = 'MoleculeModal'
 
@@ -23,4 +61,5 @@ export {
   MoleculeModalWithAnimation,
   MoleculeModalWithoutAnimation
 }
+
 export default MoleculeModalWithAnimation
